+++ news.md
@@ -1,4 +1,9 @@
 
 * rust_api: version = "0.1.0"
 
+- split the code from handlers.rs into types.rs and math.rs
+
+
+* rust_api: version = "0.1.0"
+
 rust api where you can get historical ticker data